Zwischenstand

+++ server/src/main/java/de/fhdortmund/j2t2/wise2019/server/game/remote/websocketcommands/AddMessageWebSocketCommand.java
@@ -6,16 +6,16 @@ import de.fhdortmund.j2t2.wise2019.server.commons.remote.AbstractWebSocketComman
 public class AddMessageWebSocketCommand extends AbstractWebSocketCommand {
 
     private static final String COMMAND = "AddMessage";
-    private final int chatId;
+    private final long chatId;
     private final Message message;
 
-    public AddMessageWebSocketCommand(int chatId, Message message) {
+    public AddMessageWebSocketCommand(long chatId, Message message) {
         super(COMMAND);
         this.chatId = chatId;
         this.message = message;
     }
 
-    public int getChatId() { return chatId; }
+    public long getChatId() { return chatId; }
 
     public Message getMessage() {
         return message;
